--- backend/tests/unit/apps/common/otel_test.py
@@ -1,33 +1,0 @@
-from apps.common import otel
-
-
-def test_configure_otel_metrics(mocker):
-    otel._state["configured"] = False
-    set_meter_provider = mocker.patch("apps.common.otel.metrics.set_meter_provider")
-    meter_provider = mocker.patch("apps.common.otel.MeterProvider")
-    mocker.patch("apps.common.otel.PeriodicExportingMetricReader")
-    mocker.patch("apps.common.otel.OTLPMetricExporter")
-    mocker.patch("apps.common.otel.Resource")
-    django_instrumentor = mocker.patch("apps.common.otel.DjangoInstrumentor")
-
-    otel.configure_otel_metrics()
-
-    set_meter_provider.assert_called_once_with(meter_provider.return_value)
-    django_instrumentor.return_value.instrument.assert_called_once_with(
-        meter_provider=meter_provider.return_value,
-    )
-
-
-def test_configure_otel_metrics_is_idempotent(mocker):
-    otel._state["configured"] = False
-    mocker.patch("apps.common.otel.metrics.set_meter_provider")
-    mocker.patch("apps.common.otel.MeterProvider")
-    mocker.patch("apps.common.otel.PeriodicExportingMetricReader")
-    mocker.patch("apps.common.otel.OTLPMetricExporter")
-    mocker.patch("apps.common.otel.Resource")
-    django_instrumentor = mocker.patch("apps.common.otel.DjangoInstrumentor")
-
-    otel.configure_otel_metrics()
-    otel.configure_otel_metrics()
-
-    django_instrumentor.return_value.instrument.assert_called_once()
